Round coordinates
This would have worked out of the box if this plug-in actually used the theme rather than hard-coding its sizes here. However that is too big of a change for me to make in 5 minutes, so it'll have to wait. Also fixed some missing screen scale factors.

Done as a 5 minute fix to try to amend #7125.