halve the values for meshfix_maximum_deviation
Due to a bug the setting used to be interpreted half as high as they were.
Now the bug is fixed and the values are interpreted as twice what they used to be.
In order to keep the behavior the same, we therefore halve all the values of that setting