Make support render transparently
This way you can still sort of see what is being supported under there. Handy!

However because the layer view is rendered on top of everything else, the Z buffering is a bit messed up. This makes the support always exactly the transparency value in the theme regardless of how many layers of support are drawn. That's nice, because it makes it so that you can always see through the support.

This doesn't work when the colour scheme of layer view is not set to Line Type. There is no differentiation between support and other things then. Support infill seems to not be used at all by layer view, but I've also made the support interface transparent just like the normal support.