Clarify tooltip about resolution between per-extruder values
It was using a bit of programmer jargon there that was very hard to understand. Initially I was coming in here to remove the space at the end of the translated string but I'm now changing all of it to make it more clear to the user as well.

Done as a 5 minute fix.