Don't update build volume when prime Z position is changed
It doesn't affect the build volume, since we don't need to move the prime blob disallowed area. It is projected onto the build plate after all.
Most likely this won't affect anyone since the prime Z position is a machine setting that can't normally be modified. But hey, it's a bit more precise anyway.

Found during work on the Settings Guide.