Optimise all OBJ meshes
The resolution of the coordinates in these meshes is now reduced to 1 micron. Any unnecessary trailing zeros are removed. Any unnecessary data is removed (like object name). Any comments are removed. All normals are removed and no longer used by the faces.
This should save 10MB on the download and installation size or so.