Remove code that prevents Cura from starting on MAC
The renderer isn't active when the view is activated. As such,
no opengl context *ever* exists. In some cases this might cause it to get
stuck in an indefinate loop.

The comments around it are explaining a situation that can't occur. The activateView
event of SolidView isn't triggered when switching to another view...