Skip the Cloud page if the user is logged in
If the user goes through the welcome wizard, arrives at the login page,
logs in and then closes Cura, he/she will be logged in the next time.
In this case, he/she will go through the welcome page again but will
not see the sign in page at all.

CURA-7019