Fix the welcome dialog not closing when cloud printers detected
The welcome dialog was not closing because when pressing the "Sign up"
button in the cloud page, a signal is emitted that is forcing the
welcome wizard to move to the next page. Therefore the currentPageIdx
was detected as cloud_page_idx+1.

This commit fixes the check by checking whether the welcome dialog is
in the page after the Cloud page.

CURA-7019