Fix internal conflicts within the Anycubic Mega Zero printer
This definition specified that the maximum E speed is 25mm/s but then said that the retraction speed is 30mm/s. I've changed the maximum E speed to 30mm/s then because apparently that's what the definition author was testing with.
This definition specified that the maximum Z speed is 5mm/s but the default Z hop speed is 10mm/s. I've set the Z hop speed to 5mm/s because it was not initially overridden by the definition author, but apparently 5mm/s is the limit.

Contributes to issue CURA-7431.