Rename Tronxy quality profiles to something more specific to Tronxy
To Cura the folder doesn't matter. It only reads the filenames and uses those as unique IDs. So for these filenames we'd like to have an ID that is more speciic to this printer to prevent people from crashing.
I'm guessing that the author here already encountered that problem. He copied the files from Creality which has the same problem. However Creality profiles are already published and will be much harder to change the names since they would need to be changed in a version upgrade.

Contributes to issue CURA-7384.