Fix asynchronous bug if job gets cancelled
If the print job happens to get cancelled right after checking if the index is correct, but before actually reading the line, it would get an IndexError when trying to read the line since cancelling the job clears the _gcode list.
This prevents that asynchronous issue by using the internal check in the list access to check that, and just uses an exception to check whether it's reached the end.

Fixes Sentry issue CURA-QC.