Use extruderList rather than deprecated extruders
This fixes a possible bug, too, if the extruders dictionary weren't iterated over in the order of the extruders. It would sometimes assign the profiles to the wrong extruders then. However I think the dictionary iteration in Python is in order of insertion and we probably insert them in the proper order by accident.

Done during Turbo Testing and Tooling to reduce the number of warnings in our log.