Use extruderList rather than extruders
Extruders is deprecated. Always use the extruderList.
In fact, the else case beneath was already using extruderList in the same way.
Hopefully we can one day get rid of all the string conversion for the position in this class. I don't dare to refactor that without having a review though.

Done during Turbo Testing and Tooling to reduce the number of warnings in our log.