Correctly upgrade the ironing_inset based on the ironing_pattern used
Previously, the ironing_inset was changed in the profiles to reflect
the changes made in the CuraEngine. This inset though was not taking
into account the ironing_pattern (in the CuraEngine, the inset is
compensated differently depending on the pattern).

This commit fixes that in the version upgrader and the printer
definitions in order to maintain the same behavior for the old
profiles.

CURA-7318