Avoid hitting the prime blob when Z hop is enabled
If Z hop is enabled, it would move to the height of the Z hop at the start of the print, in order to hop for the initial travel move. Cura sort of assumes that this hop is going up and so it's safe to do, but in reality this hop is going down due to the starting process in the firmware. As a hotfix, we'll just move away from that blob so that we don't hit it even if you enable Z hops.