Run CTest in parallel (#7952)
* Don't pre-load engine before other plug-ins

This was a fix for the post-processing plug-in that's no longer required.
It was also only applied when running from source, so obviously it's not required or we'd have seen it in the builds.

* Run automated tests in parallel

Done during Turbo Testing and Tooling.

* Pass number of processors via ctest command itself

More reliable apparently because it's a CMake variable that doesn't get cached. It shouldn't get cached or set in the CMake script because it's not the same for every computer.

Done during Turbo Testing and Tooling.