Apply transformation inside 3MF file
It turns out that we have another bug in Uranium: Transformations from 3MF files are not applied to the platform mesh. This makes sense because UM.Scene.Platform takes the mesh data (without transformations) from the node that it read and squashes that into the SceneNode for the platform. This SceneNode already has a transformation for the platform offset. The 3MF node also has a transformation from the 3MF's convention to have the origin in the front left. The 3MF reader then squishes the transformation from the node into that.
To fix this we'd need to either:
* separate the transformation of the 3MF file from the transformation from the 3MF convention to have the origin in the front left, so that we can pick and choose, or
* remove the transformation from the 3MF convention and apply it only on loading normal printed meshes, and then modify all of the 3MF platform files.

Both require significant effort. So instead I applied the transformation directly to the vertex data.
To do that I imported the file into Blender using my own Blender plug-in, exported to STL and imported that STL again to apply the transformations, and then exported the file from Blender again.
In Blender I also combined a lot of flat faces together, which reduced the file type losslessly.

Contributes to issue CURA-7534.