Remove redundant button.
As discussed in the CCB this morning. Since the new profile will be the _changed to_ profile rather than the _changed from_ (the latter of which would make the button useful, but since the profile is already switched at that point it would be a lot of work), the 'Create New Profile ...' button (in  the Keep/Discard dialog) is redundant, since you can always create a new profile if you select keep changes.