Fix visibility of some settings regarding support structure and support meshes
I think for the support meshes, some settings were simply missed because they are in completely different categories. For instance, the support flow rate also applies to support meshes so those settings should be visible.

For the tree support settings, I think that when the tree support became activated by support_structure, the developer checked for instances where the support_tree_enable setting made a setting visible (as alternative to the normal support_enable setting). Like Support Infill Density used to be visible if normal support or tree support was enabled, and now needs to check only for normal support. However they didn't also look for settings that are ONLY available for normal support and not for tree support. So those settings now have extra conditions that disable them if the support structure is set to tree support.
With this I tried to word the check correctly in case we ever get a third support construction type. Like, sometimes a setting is not shown with tree support because tree support explicitly doesn't support it (e.g. Conical Support), but sometimes it's not shown with tree support because it's specific to area support (e.g. support join distance).

Done as a five minute fix (although typing this commit message also took five minutes) as I encountered these problems while working on the settings guide.