Document possible theoretical issue when loading a project file
The issue will happen only if the machine_extruder_count is a formula.
In this case, the project loading will not work properly if "Create new" printer is selected,
as the settings of all extruders but the first one will not be applied.
Workaround in this case will be to load the project again and select to update the existing
printer, in which case all settings will load properly.

CURA-7646