Also show Connect Support Lines for lines pattern
Previously this was disabled because you should probably use the ZigZag pattern then. But I'm enabling it now because of two inconsistencies:
* The setting still works. So you can change the setting by switching to grid, triangles, cross or gyroid, then switch back to lines and still get connected support lines.
* The same happens with the infill patterns, where this setting is visible for the lines pattern.

Rather than changing the other two things (which involves changes to CuraEngine too to make it ignore the setting value for the lines pattern), I've opted to change it here and fix both inconsistencies at once. If I were to make the setting dysfunctional, it would also lead to user confusion, because they then need to discover that they need to change the infill pattern, which may look unlogical to them.

Fixes #8192.