Change label and resolve functionality of shrinkage compensation setting
We'll now actually compensate for the shrinkage. That means that the setting needs to become user-visible. We need to make that experience a bit more user-friendly.

The setting is now also global, with a resolve function. That means that the user can't change this per extruder any more, because the scale factor applied in the engine can only be global. The resolve function means that profiles can still apply a per-extruder value but this value gets resolved with other extruders' values.

Contributes to issue CURA-7118.