Adding the new Digital Factory URL instead of the old one
If the user is already logged in, we used to direct him to the https://mycloud.ultimaker.com page. Besides the fact that this URL is going to be phased out, there will be problems whenever there is a staging build, since it will always point the user to the production environment.