Assign FABtotum profiles to 0.4mm nozzle
This is the default (preferred) nozzle for the FABtotum Personal Fabricator.
Apparently when these nozzle profiles were added, the existing quality profiles weren't considered. We'll now put them on the nozzle size that the FABtotum printer had before the nozzle profiles were added. There is also the Pro nozzle for FABtotum but I think the Lite is more likely what the profiles were originally made for, since that is the default.