Check for updatable machines after the machine_definition_id changes
So that we can account for the issue with the Dagoma printers (CURA-7517), which when they were
updated, the dagoma_discoeasy200 (dual extruder) was renamed to dagoma_discoeasy200_bicolor
and the dagoma_discoeasy200 became single extrusion. In this case, the project files created
before 4.6 should look for the dual extrusion dagomas (dagoma_discoeasy200_bicolor).

CURA-7644