Remove "Enable Variable Line Widening" in favor of "Print Thin Walls"
The setting "Enable Variable Line Widening" was using the same notion as the Print Thin Walls.
This commit fixes that by removing it and making the settings "Minimum Feature Size" and "Minimum
Wall Line Width" dependand on the Print Thin Walls.

The commit also adds reasonable default and minimum/maximum values for those two settings, along
with a clearer description on how these settings affect the thin features of the model.

CURA-7682