Reduce minimum value and default value for minimum line width
The previous minimum value was based on anecdotal experience about what the minimum width is that a printer can normally print. That's better as a default value and shouldn't be used as a minimum value.
The minimum value should be based on what the technical or physical limitations are. There are no real technical or physical limitations other than that you can't have 0-width lines. So the minimum is just-above-0.

Contributes to issue CURA-7682.