Don't create DiscardOrKeepProfileChangesDialog on start-up
Just whenever we need it, please.
This removes a warning from the start-up sequence, but the warning is still there when you open the dialogue. It's complaining that a model has size -13 and another has size -14, but I don't know which model that is. The UserChangesModel seems to be fine at a size of 1 or so.

Contributes to issue CURA-7501.