Fix all places that one-at-a-time is affecting
When changing the scene based on one-at-a-time mode, it should also be checked if the global stack
has only one extruder enabled. If one-at-a-time is enabled, but there are more than one extruders
enabled, the one-at-a-time mode is no longer valid and it can be inferred that the objects will be
printed all at once.

**Note:** If you enable one-at-a-time and then enable the second extruder, the global stack will
still report that it has one-at-a-time print sequence, despite the fact that it will print
everything all at once, since the setting is not resetted to all-at-once (compared what the
original fix https://github.com/Ultimaker/Cura/commit/9266c753c30868ca3b6badda41da1b0f6d9b9857 was attempting to do.

CURA-7827